move functions to test-driver
